serverless_data_validation_GE/requirements.txt
Flask==2.0.2
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.1.0
google-cloud-bigquery-storage==2.10.1
pybigquery==0.5.0
google-cloud-storage==1.44.0
google-api-python-client==2.34.0
sqlalchemy==1.3.24
pyarrow==6.0.1
great-expectations==0.14.0